raft: make election and heartbeat ticks of type int64
In the near future, we want to keep incrementing the electionElapsed
until we hear from a leader. In order to avoid any chances of an integer
overflow, this commit changes the electionElapsed (along with other
symmetrical variables from int to int64)

Epic: None

Release note: None

+++ pkg/kv/kvserver/replica.go
@@ -811,10 +811,10 @@ type Replica struct {
 		quotaReleaseQueue []*quotapool.IntAlloc
 
 		// Counts calls to Replica.tick()
-		ticks int
+		ticks int64
 
 		// lastProposalAtTicks tracks the time of the last proposal, in ticks.
-		lastProposalAtTicks int
+		lastProposalAtTicks int64
 
 		// Counts Raft messages refused due to queue congestion.
 		droppedMessages int
